add devise, skeleton for manage backend, permissions, and polish map UI

+++ app/controllers/tags_controller.rb
@@ -1,6 +1,7 @@
 class TagsController < ApplicationController
   def destroy
     @tag = Tag.find(params[:id])
+    authorize @tag
 
     if @tag.destroy
       render :destroy
